Old implementation

+++ core/codenvy-hosted-api-organization/src/main/java/com/codenvy/api/resources/model/Resource.java
@@ -27,7 +27,7 @@ public interface Resource {
     /**
      * Returns amount of resources
      */
-    int getAmount();
+    long getAmount();
 
     /**
      * Returns unit of resources, e.g. for RAM type it should be equal to mb
